refactor: reduce directories
