+++ misc/text_chart_visualization.py
